wip: bigwig refact
