+++ src/main/data/formats/bbi.js
@@ -6,35 +6,91 @@
 
 'use strict';
 
+import _ from 'underscore';
+
 import {typeAtOffset} from './helpers';
 
-var TYPE_SET = {
+var BigBedHeader = {
+  _magic: ['const', 'uint32', 0x8789F2EB, true],
+  version: 'uint16',
+  zoomLevels: ['const', 'uint16', 0 ],
+  chromosomeTreeOffset: 'uint64',
+  unzoomedDataOffset: 'uint64',
+  unzoomedIndexOffset: 'uint64',
+  fieldCount: 'uint16',
+  definedFieldCount: 'uint16',
+  // 0 if no autoSql information
+  autoSqlOffset: 'uint64',
+  totalSummaryOffset: 'uint64',  // TODO: restrict to version ≥ 2. Is it relevant at all in BigBed?
+  // Size of uncompression buffer.  0 if uncompressed.
+  uncompressBufSize: 'uint32',
+  // Offset to header extension 0 if no such extension
+  // TODO: support extended headers (not used in ensGene.bb); spec calls this block simply 'reserved'.
+  extensionOffset: 'uint64',
+  zoomHeaders: ['array', 'ZoomHeader', 'zoomLevels'],
+
+  totalSummary: typeAtOffset('TotalSummary', 'totalSummaryOffset'),
+  chromosomeTree: typeAtOffset('BPlusTree', 'chromosomeTreeOffset')
+};
+
+var BigWigHeader = _.extend({}, BigBedHeader);
+BigWigHeader.magic = [ 'const', 'uint32', 0x888FFC26, true ];
+BigWigHeader.zoomLevels = 'uint16';
+BigWigHeader.fieldCount = [ 'const', 'uint16', 0, true ];
+BigWigHeader.definedFieldCount = [ 'const', 'uint16', 0, true ];
+BigWigHeader.autoSqlOffset = [ 'const', 'uint64', 0, true ];
+
+var CirTree = {
   'jBinary.littleEndian': true,
 
-  'Header': {
-    _magic: ['const', 'uint32', 0x8789F2EB, true],
-    version: 'uint16',
-    zoomLevels: 'uint16',
-    chromosomeTreeOffset: 'uint64',
-    unzoomedDataOffset: 'uint64',
-    unzoomedIndexOffset: 'uint64',
-    fieldCount: 'uint16',
-    definedFieldCount: 'uint16',
-    // 0 if no autoSql information
-    autoSqlOffset: 'uint64',
-    totalSummaryOffset: 'uint64',
-    // Size of uncompression buffer.  0 if uncompressed.
-    uncompressBufSize: 'uint32',
-    // Offset to header extension 0 if no such extension
-    // TODO: support extended headers (not used in ensGene.bb)
-    extensionOffset: 'uint64',
-    zoomHeaders: ['array', 'ZoomHeader', 'zoomLevels'],
-
-    totalSummary: typeAtOffset('TotalSummary', 'totalSummaryOffset'),
-    chromosomeTree: typeAtOffset('BPlusTree', 'chromosomeTreeOffset')
+  // The "CIR" tree contains a mapping from sequence -> block offsets; it stands for "Chromosome Index R tree".
+  CirTree: {
+    _magic: [ 'const', 'uint32', 0x2468ACE0, true ],
+    blockSize: 'uint32',
+    itemCount: 'uint64',
+    startChromIx: 'uint32',
+    startBase: 'uint32',
+    endChromIx: 'uint32',
+    endBase: 'uint32',
+    fileSize: 'uint64',
+    itemsPerSlot: 'uint32',
+    _reserved: ['skip', 4],
+    blocks: 'CirNode'
   },
 
-  'TotalSummary': {
+  CirNode: {
+    isLeaf: 'uint8',  // 1 = yes, 0 = no
+    _reserved: 'uint8',
+    count: 'uint16',
+    contents: [
+      'array',
+      [ 'if', 'isLeaf', 'LeafData', 'NonLeafData' ],
+      'count'
+    ]
+  },
+
+  LeafData: {
+    startChromIx: 'uint32',
+    startBase: 'uint32',
+    endChromIx: 'uint32',
+    endBase: 'uint32',
+    offset: 'uint64',
+    size: 'uint64'
+  },
+
+  NonLeafData: {
+    startChromIx: 'uint32',
+    startBase: 'uint32',
+    endChromIx: 'uint32',
+    endBase: 'uint32',
+    offset: 'uint64'
+  },
+};
+
+var BigBedWigTypeSet = {
+  'jBinary.littleEndian': true,
+
+  TotalSummary: {
     basesCovered: 'uint64',
     minVal: 'float64',     // for bigBed minimum depth of coverage
     maxVal: 'float64',     // for bigBed maximum depth of coverage
@@ -42,14 +98,7 @@ var TYPE_SET = {
     sumSquared: 'float64'  // for bigBed sum of coverage squared
   },
 
-  'ZoomHeader': {
-    reductionLevel: 'uint32',
-    _reserved: 'uint32',
-    dataOffset: 'uint64',
-    indexOffset: 'uint64'
-  },
-
-  'BPlusTree': {
+  BPlusTree: {
     magic: ['const', 'uint32', 0x78CA8C91, true],
     // Number of children per block (not byte size of block)
     blockSize: 'uint32',
@@ -63,8 +112,8 @@ var TYPE_SET = {
     _reserved3: ['skip', 4],
     nodes: 'BPlusTreeNode'  // ['array', 'BPlusTreeNode', 'itemCount']
   },
-  
-  'BPlusTreeNode': {
+
+  BPlusTreeNode: {
     isLeaf: 'uint8',  // 1 = yes, 0 = no
     _reserved: 'uint8',
     count: 'uint16',
@@ -80,118 +129,111 @@ var TYPE_SET = {
     }], 'count']
   },
 
-  'CirTree': {
-    _magic: ['const', 'uint32', 0x2468ACE0, true],
-    blockSize: 'uint32',
-    itemCount: 'uint64',
-    startChromIx: 'uint32',
-    startBase: 'uint32',
-    endChromIx: 'uint32',
-    endBase: 'uint32',
-    fileSize: 'uint64',
-    itemsPerSlot: 'uint32',
-    _reserved: ['skip', 4],
-    blocks: 'CirNode'
+  ZoomLevel: {
+    count: 'uint32',
+    data: ['array', 'ZoomData', 'count']
   },
 
-  'CirNode': {
-    isLeaf: 'uint8',  // 1 = yes, 0 = no
-    _reserved: 'uint8',
-    count: 'uint16',
-    contents: [
-      'array',
-      [ 'if', 'isLeaf', 'LeafData', 'NonLeafData' ],
-      'count'
-    ]
+  ZoomData: {
+    chrId: 'uint32',
+    start: 'uint32',
+    end: 'uint32',
+    validCount: 'uint32',
+    minVal: 'uint32',
+    maxVal: 'uint32',
+    sum: 'uint32',
+    sumSqs: 'uint32'
   },
-
-  'LeafData': {
-    startChromIx: 'uint32',
-    startBase: 'uint32',
-    endChromIx: 'uint32',
-    endBase: 'uint32',
-    offset: 'uint64',
-    size: 'uint64'
-  },
-
-  'NonLeafData': {
-    startChromIx: 'uint32',
-    startBase: 'uint32',
-    endChromIx: 'uint32',
-    endBase: 'uint32',
-    offset: 'uint64'
-  },
-  
-  'ZoomLevel': {
-    'count': 'uint32',
-    'data': [ 'array', 'ZoomData', 'count' ]
-  },
-
-  'ZoomData': {
-    'chrId': 'uint32',
-    'start': 'uint32',
-    'end': 'uint32',
-    'validCount': 'uint32',
-    'minVal': 'uint32',
-    'maxVal': 'uint32',
-    'sum': 'uint32',
-    'sumSqs': 'uint32'
-  },
-
-  'WigData': {
-    'chrId': 'uint32',
-    'start': 'uint32',
-    'stop': 'uint32',
-    'step': 'uint32',
-    'span': 'uint32',
-    'tpe': 'uint8',
-    'reserved': 'uint8',
-    'count': 'uint16',
-    'data': [
-      'array',
-      [
-        'if',
-        (ctx) => { return ctx.tpe == 1; },
-        'WigFixedData',
-        [
-          'if',
-          (ctx) => { return ctx.tpe == 2; },
-          'WigVarData',
-          [
-            'if',
-            (ctx) => { return ctx.tpe == 3; },
-            'WigBedGraphData',
-            'WigBedGraphData'  // TODO(ryan): should be an error…
-          ]
-        ]
-      ],
-      'count'
-    ]
-  },
-  
-  'WigFixedData': {
-    'value': 'float32'
-  },
-
-  'WigVarData': {
-    'start': 'uint32',
-    'value': 'float32'
-  },
-
-  'WigBedGraphData': {
-    'start': 'uint32',
-    'end': 'uint32',
-    'value': 'float32'
-  },
-
-  'BedEntry': {
-    'chrId': 'uint32',
-    'start': 'uint32',
-    'stop': 'uint32',
-    'rest': 'string0'
-  },
-
-  'BedBlock': ['array', 'BedEntry'],
 };
 
-module.exports = {TYPE_SET};
+var BigBedTypeSet = _.extend(
+  {},
+  BigBedWigTypeSet
+);
+BigBedTypeSet.Header = BigBedHeader;
+
+var BedBlockTypeSet = {
+  'jBinary.littleEndian': true,
+
+  Entry: {
+    chrId: 'uint32',
+    start: 'uint32',
+    stop: 'uint32',
+    rest: 'string0'
+  },
+
+  Block: [ 'array', 'Entry' ]
+};
+
+var BigWigTypeSet = _.extend(
+  {
+    Header: BigWigHeader,
+
+    ZoomHeader: {
+      reductionLevel: 'uint32',
+      _reserved: 'uint32',
+      dataOffset: 'uint64',
+      indexOffset: 'uint64'
+    },
+
+    FixedData: {
+      value: 'float32'
+    },
+
+    VarData: {
+      start: 'uint32',
+      value: 'float32'
+    },
+
+    BedGraphData: {
+      start: 'uint32',
+      end: 'uint32',
+      value: 'float32'
+    },
+
+    Data: {
+      chrId: 'uint32',
+      start: 'uint32',
+      stop: 'uint32',
+      step: 'uint32',
+      span: 'uint32',
+      tpe: 'uint8',
+      reserved: 'uint8',
+      count: 'uint16',
+      data: [
+        'array',
+        [
+          'if',
+          (ctx) => {
+            return ctx.tpe == 1;
+          },
+          'FixedData',
+          [
+            'if',
+            (ctx) => {
+              return ctx.tpe == 2;
+            },
+            'VarData',
+            [
+              'if',
+              (ctx) => {
+                return ctx.tpe == 3;
+              },
+              'BedGraphData',
+              'BedGraphData'  // TODO(ryan): should be an error…
+            ]
+          ]
+        ],
+        'count'
+      ]
+    },
+  },
+  BigBedWigTypeSet
+);
+
+module.exports = {
+  BigBedTypeSet,
+  BigWigTypeSet,
+  BedBlockTypeSet,
+  CirTree
+};
